Fix includedir
noson-app code uses `<noson/*.h>`, so building that against noson as
system library will fail as `<noson/noson/*.h>` does not exist.

This is the case on OpenBSD where /usr/local/ is the default prefix.
On Linux, the default is /usr/, i.e. the same as the default prefix.

My guess it's been working by accident so far.